refactor: do not use `WidgetModel`, `DataTableModel`
